spec: openflow.h: ethertype is 16 bits

+++ include/openflow/openflow.h
@@ -497,7 +497,7 @@ OFP_ASSERT(sizeof(struct ofp_action_mpls_ttl) == 8);
 struct ofp_action_push {
     uint16_t type;                  /* OFPAT_PUSH_VLAN/MPLS. */
     uint16_t len;                   /* Length is 8. */
-    uint8_t ethertype;              /* Ethertype */
+    uint16_t ethertype;             /* Ethertype */
     uint8_t pad[2];
 };
 OFP_ASSERT(sizeof(struct ofp_action_push) == 8);
@@ -506,7 +506,7 @@ OFP_ASSERT(sizeof(struct ofp_action_push) == 8);
 struct ofp_action_pop_mpls {
     uint16_t type;                  /* OFPAT_POP_MPLS. */
     uint16_t len;                   /* Length is 8. */
-    uint8_t ethertype;              /* Ethertype */
+    uint16_t ethertype;             /* Ethertype */
     uint8_t pad[2];
 };
 OFP_ASSERT(sizeof(struct ofp_action_pop_mpls) == 8);
